backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymongo==4.6.1
motor==3.3.2

# AI & Machine Learning
openai==1.10.0
google-generativeai==0.8.3
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Caching & Performance
redis==5.0.1
asyncio==3.4.3

# API Documentation
swagger-ui-bundle==0.0.9

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
